Rename openguessit to guessit
