+++ app/polochon_modules.go
@@ -7,9 +7,9 @@ import (
 	_ "github.com/odwrtw/polochon/modules/canape"
 	_ "github.com/odwrtw/polochon/modules/eztv"
 	_ "github.com/odwrtw/polochon/modules/fsnotify"
+	_ "github.com/odwrtw/polochon/modules/guessit"
 	_ "github.com/odwrtw/polochon/modules/imdb"
 	_ "github.com/odwrtw/polochon/modules/mock"
-	_ "github.com/odwrtw/polochon/modules/openguessit"
 	_ "github.com/odwrtw/polochon/modules/opensubtitles"
 	_ "github.com/odwrtw/polochon/modules/pushover"
 	_ "github.com/odwrtw/polochon/modules/tmdb"
